Eliminate arbitrary multiplier and calc using double floats instead
N.B! DB Change:
- cares_impactmatrix has new type (double precision[][])

+++ symphony-ws/src/main/java/se/havochvatten/symphony/calculation/SankeyChart.java
@@ -47,7 +47,7 @@ public class SankeyChart {
         return pLen + e; // TODO do something more clever?
     }
 
-    public SankeyChart(int[] ecosystemServices, int[] pressures, long[][] impactMatrix, double total)
+    public SankeyChart(int[] ecosystemServices, int[] pressures, double[][] impactMatrix, double total)
     {
         int pLen = impactMatrix.length, esLen = impactMatrix[0].length;
